fix(ci): line-limit — extract retry runtime helpers and closed adoption tests
Keep workspaces_retry.py and hosted adoption tests under the 1500-line
first-party gate after restoring the closed planning-scope replacement path.
