+++ tests/unit/service/test_workspace_retry_pr_lifecycle.py
@@ -7,6 +7,7 @@ from types import SimpleNamespace
 import pytest
 
 import awf.service.workspaces_retry as workspaces_retry_service
+import awf.service.workspaces_retry_runtime as workspaces_retry_runtime
 from awf.common.forge_lifecycle import PullRequestLifecycle
 from awf.service.workspaces import WorkspaceRetryNotFoundError
 from awf.service.workspaces_retry import _live_pr_lifecycle, _pr_number_from_url
@@ -49,7 +50,7 @@ async def test_live_pr_lifecycle_uses_current_forge_status(
             return PullRequestLifecycle.merged
 
     monkeypatch.setattr(
-        workspaces_retry_service,
+        workspaces_retry_runtime,
         "make_forge_client",
         lambda _forge, _runner: FakeForgeClient(),
     )
